% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ei_df.R
\name{as.ei_df}
\alias{as.ei_df}
\alias{as.ei_df}
\alias{as.ei_df.data.frame}
\alias{as.ei_df}
\alias{as.ei_df.ei_df}
\title{Convert foreign object to an \code{ei_df} object}
\usage{
as.ei_df(x, ...)

\method{as.ei_df}{data.frame}(x, ..., intervals, energy, water, ei)

\method{as.ei_df}{ei_df}(x, ...)
}
\arguments{
\item{x}{object to be converted into an object class \code{ei_df}}

\item{...}{passed on to \link{ei_df}, might included \code{method}}

\item{energy}{names or numbers of the numeric columns holding energy data}

\item{water}{names or numbers of the numeric columns holding water data}

\item{ei}{name or number of the numeric columns holding ei data}

\item{interval}{names or numbers of the \code{interval} column}
}
\description{
Convert foreign object to an \code{ei_df} object
}
\examples{
library(units)
x <- data.frame(water = set_units(1:10, mg),
                energy = set_units(rnorm(10), kWh))
as.ei_df(x)
}
