% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ei_df.R
\name{ei_df}
\alias{ei_df}
\title{Create \code{ei_df} object}
\usage{
ei_df(..., stringsAsFactors = default.stringsAsFactors())
}
\arguments{
\item{...}{column elements to be binded into an \code{ei_df} object or a single \code{list} or \code{data.frame} with required columns.}

\item{stringsAsFactors}{logical; logical: should character vectors be converted to factors?  The `factory-fresh' default is \code{TRUE}, but this can be changed by setting \code{options(stringsAsFactors = FALSE)}.}
}
\description{
Create \code{ei_df}, which extends data.frame-like objects with standardized energy intensity data series
}
\examples{
# ei_df()
}
