% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/energyintensity.R
\docType{package}
\name{energyintensity}
\alias{energyintensity}
\alias{energyintensity-package}
\alias{energyintensity-package}
\title{energyintensity: water system embedded energy modeling}
\description{
Network-based modeling of embedded energy (energy intensity) within a water distribution system
}
\section{Classes}{

The ei classes ...
}

\section{Functions}{

The ei functions ...
}

\author{
\strong{Maintainer}: Andrew Holguin \email{ajholguin@ucdavis.edu}

}
