% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meter_df.R
\name{meter_df}
\alias{meter_df}
\title{Create \code{meter_df} object}
\usage{
meter_df(..., stringsAsFactors = default.stringsAsFactors())
}
\arguments{
\item{...}{column elements to be binded into an \code{meter_df} object or a single \code{list} or \code{data.frame} with required columns.}

\item{stringsAsFactors}{logical; logical: should character vectors be converted to factors?  The `factory-fresh' default is \code{TRUE}, but this can be changed by setting \code{options(stringsAsFactors = FALSE)}.}
}
\description{
Create \code{meter_df}, which extends data.frame-like objects with standardized meter data series
}
\examples{
# meter_df()
}
