% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{pst1_energy}
\alias{pst1_energy}
\title{Sample Energy Meter data}
\format{A data frame with 53940 rows and 10 variables:
\describe{
  \item{interval}{time interval}
  \item{energy}{kWh}
}}
\usage{
pst1_energy
}
\description{
Simulated data, designed to mimic typical energy meter data.
}
\keyword{datasets}
