% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{pst1_water}
\alias{pst1_water}
\title{Sample Water Meter data}
\format{A data frame with 53940 rows and 10 variables:
\describe{
  \item{interval}{time interval}
  \item{water}{mg}
}}
\usage{
pst1_water
}
\description{
Simulated data, designed to mimic typical water meter data.
}
\keyword{datasets}
