% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reconcile.R
\name{reconcile}
\alias{reconcile}
\title{Merge and Reconcile Meter Data}
\usage{
reconcile(...)
}
\arguments{
\item{...}{\code{meter_df} objects, which will be merged and reconciled}
}
\description{
Merge multiple \code{meter_df} objects together so that the values can be
compared. This often requires aggregation or resampling of the original
values.
}
\examples{
library("units")

x <- data.frame(time = seq.Date(as.Date("2013-1-1"), by = "+1 month", length.out = 12),
                value = set_units(rnorm(12), kW))

meter_df(x)
}
