% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meter_df.R
\name{meter_df}
\alias{meter_df}
\title{Create \code{meter_df} object}
\usage{
meter_df(..., stringsAsFactors = default.stringsAsFactors())
}
\arguments{
\item{...}{column elements to be binded into an \code{meter_df} object or a \code{data.frame} with required columns. One element must be a time index and at least one should have measured values with units set, see \code{set_units}}

\item{stringsAsFactors}{logical; logical: should character vectors be converted to factors?  The `factory-fresh' default is \code{TRUE}, but this can be changed by setting \code{options(stringsAsFactors = FALSE)}.}
}
\description{
Create \code{meter_df}, which is essentially a data.frame with standardized meter data series
}
\examples{
library("units")

x <- data.frame(time = seq.Date(as.Date("2013-1-1"), by = "+1 month", length.out = 12),
                value = set_units(rnorm(12), kW))

meter_df(x)
}
